backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
passlib[bcrypt]
python-dotenv
alembic
psycopg2-binary
gunicorn
psycopg2-binary
